+++ index.html
@@ -56,6 +56,6 @@
 		<script src="node_modules/jquery/dist/jquery.js"></script>
 		<script src="node_modules/handlebars/dist/handlebars.js"></script>
 		<script src="node_modules/director/build/director.js"></script>
-		<script src="js/app.ts"></script>
+		<script src="dist/app.js"></script>
 	</body>
 </html>
