+++ js/util/pluralize.ts
@@ -1,0 +1,4 @@
+const pluralize = (count, word) => {
+  return count === 1 ? word : word + 's';
+}
+export default pluralize
