+++ js/util/store.ts
@@ -1,0 +1,10 @@
+const store = function(namespace, data?) {
+  if (arguments.length > 1) {
+    return localStorage.setItem(namespace, JSON.stringify(data));
+  } else {
+    var store = localStorage.getItem(namespace);
+    return (store && JSON.parse(store)) || [];
+  }
+}
+
+export default store
