+++ js/util/util.ts
@@ -1,0 +1,11 @@
+import uuid from './uuid';
+import pluralize from './pluralize';
+import store from './store';
+
+const util = {
+  uuid,
+  pluralize,
+  store
+}
+
+export default util
