+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@types/handlebars": "^3.0.28",
     "@types/jquery": "^2.0.54",
+    "axios": "^0.19.0",
     "director": "~1.2.2",
     "handlebars": "~2.0.0",
     "jquery": "~2.1.1",
